Disable Ancestor persistence between saves
This is because I keep seeing errors on save game load, and they're not
immediately obvious. I *think* what's happening is I push them into the
list and then one of them dies, because the first misaligned load error
comes up says that a Corpse is being read as a Building - but I really
don't know enough to say!

So I'm kicking the can down the road.