Fix bug with spawners count
I turned the spawners storage from a set into a list some time ago to
avoid having to save the current spawner, but the call that added items
into the list was called on item spawn.

So, when you load the game and it spawned another shrine it would push
another ref to itself onto the list!

Fixed to push only after Make (construction)